+++ ichnaea/api/locate/schema.py
@@ -20,6 +20,7 @@ from ichnaea.models.cell import (
 from ichnaea.models import constants
 from ichnaea.models.constants import Radio
 from ichnaea.models.mac import (
+    channel_frequency,
     encode_mac,
     MacNode,
 )
@@ -358,6 +359,12 @@ class ValidWifiLookupSchema(colander.MappingSchema, ValidatorNode):
         validator=colander.Range(
             constants.MIN_WIFI_CHANNEL, constants.MAX_WIFI_CHANNEL))
 
+    frequency = DefaultNode(
+        colander.Integer(),
+        missing=None,
+        validator=colander.Range(
+            constants.MIN_WIFI_FREQUENCY, constants.MAX_WIFI_FREQUENCY))
+
     signalStrength = DefaultNode(
         colander.Integer(),
         missing=None,
@@ -371,33 +378,18 @@ class ValidWifiLookupSchema(colander.MappingSchema, ValidatorNode):
             constants.MIN_WIFI_SNR, constants.MAX_WIFI_SNR))
 
     def deserialize(self, data):
-        if data:
+        data = super(ValidWifiLookupSchema, self).deserialize(data)
+        if (data and data is not colander.drop and data is not colander.null):
             channel = data.get('channel')
-            channel = channel is not None and int(channel) or None
-
-            if (channel is None or not
-                    (constants.MIN_WIFI_CHANNEL <= channel <=
-                     constants.MAX_WIFI_CHANNEL)):
+            frequency = data.get('frequency')
+            if ((frequency is None and channel is not None) or
+                    (frequency is not None and channel is None)):
                 # shallow copy
                 data = dict(data)
+                data['channel'], data['frequency'] = channel_frequency(
+                    channel, frequency)
 
-                # if no explicit channel was given, calculate
-                freq = data.get('frequency', None)
-                if freq is None:
-                    freq = 0
-
-                if 2411 < freq < 2473:
-                    # 2.4 GHz band
-                    data['channel'] = (freq - 2407) // 5
-                elif freq == 2484:
-                    data['channel'] = 14
-                elif 5169 < freq < 5826:
-                    # 5 GHz band
-                    data['channel'] = (freq - 5000) // 5
-                else:
-                    data['channel'] = None
-
-        return super(ValidWifiLookupSchema, self).deserialize(data)
+        return data
 
 
 class WifiLookup(BaseLookup):
@@ -408,6 +400,7 @@ class WifiLookup(BaseLookup):
         'macAddress',
         'age',
         'channel',
+        'frequency',
         'signalStrength',
         'signalToNoiseRatio',
         'ssid',